improvement(redux): refactored with redux-ducks
