+++ src/components/PrivateRoute/PrivateRoute.tsx
@@ -1,13 +1,14 @@
 import React from 'react';
 import { Route, Redirect } from 'react-router-dom';
 import { connect } from 'react-redux';
+import { getSignIn } from '../../redux/signIn';
 
 const mapStateToProps = (state: any) => ({
-  isAuthenticated: state.signIn.isAuthenticated,
+  signIn: getSignIn(state),
 });
 
 const PrivateRoute: React.FC<any> = (props) => {
-  const { component: Component, isAuthenticated, ...restProps } = props;
+  const { component: Component, signIn: { isAuthenticated }, ...restProps } = props;
 
   return (
     <Route {...restProps}
