+++ src/redux/signIn/actions.ts
@@ -1,0 +1,6 @@
+import { createAction } from 'redux-actions';
+
+export const signInRequest = createAction('SIGN_IN_REQUEST');
+export const signInSuccess = createAction('SIGN_IN_SUCCESS');
+export const signInFailure = createAction('SIGN_IN_FAILURE');
+export const logout = createAction('SIGN_OUT');
