+++ src/redux/signIn/selectors.ts
@@ -1,0 +1,1 @@
+export const getSignIn = (state: any) => state.signIn;
