+++ src/redux/signInForm/actions.ts
@@ -1,0 +1,3 @@
+import { createAction } from 'redux-actions';
+
+export const createSignInData = createAction('CREATE_SIGN_IN_DATA');
