+++ src/redux/signUp/actions.ts
@@ -1,0 +1,5 @@
+import { createAction } from 'redux-actions';
+
+export const signUpRequest = createAction('SIGN_UP_REQUEST');
+export const signUpSuccess = createAction('SIGN_UP_SUCCESS');
+export const signUpFailure = createAction('SIGN_UP_FAILURE');
