+++ src/redux/signUp/selectors.ts
@@ -1,0 +1,1 @@
+export const getSignUp = (state: any) => state.signUp;
