+++ src/redux/signUpForm/actions.ts
@@ -1,0 +1,3 @@
+import { createAction } from 'redux-actions';
+
+export const createSignUpData = createAction('CREATE_SIGN_UP_DATA');
