+++ src/redux/signUpForm/index.ts
@@ -1,0 +1,2 @@
+export { default } from './reducer';
+export * from './actions';
