+++ src/redux/userCardForm/actions.ts
@@ -1,0 +1,3 @@
+import { createAction } from 'redux-actions';
+
+export const createUserCardData = createAction('CREATE_USER_CARD_DATA');
